+++ extension/customize/package.json
@@ -4,7 +4,7 @@
   "scripts": {
     "dev": "next dev",
     "lint": "next lint",
-    "build": "next build && next export && mv out/_next out/assets && sed -i '' -e 's/\\/_next/\\.\\/out\\/assets/g' out/**.html && mv out/index.html out/popup.html && mv out/popup.html ../ && mv out ../",
+    "build": "next build && next export && mv out/_next out/next && sed -i '' -e 's/\\/_next/\\.\\/customize\\/out\\/next/g' out/**.html && cp out/index.html ../customize.html",
     "start": "next start",
     "check:prettier": "npx prettier --check .",
     "write:prettier": "npx prettier --write ."
