+++ extension/manifest.v2.json
@@ -1,9 +1,9 @@
 {
   "name": "Minimal Theme for Twitter",
-  "short_name": "Minimal Twitter",
-  "description": "Declutters the Twitter web experience.",
+  "short_name": "MinimalTwitter",
+  "description": "Declutter your Twitter web experience.",
   "version": "4.0.0",
-  "manifest_version": 2,
+  "manifest_version": 3,
   "icons": {
     "16": "images/MinimalTwitterIcon16.png",
     "32": "images/MinimalTwitterIcon32.png",
@@ -15,12 +15,12 @@
     {
       "run_at": "document_start",
       "matches": ["https://twitter.com/*", "https://mobile.twitter.com/*"],
-      "css": ["content.css"],
-      "js": ["content.js"]
+      "css": ["content/main.css"],
+      "js": ["content/main.js"]
     }
   ],
   "options_ui": {
-    "page": "index.html",
+    "page": "customize.html",
     "open_in_tab": true
   },
   "browser_action": {
@@ -30,6 +30,6 @@
       "48": "images/MinimalTwitterIcon48.png"
     },
     "default_title": "Customize",
-    "default_popup": "index.html"
+    "default_popup": "customize.html"
   }
 }
